Moved add new card form to fragment. Added focus listeners and text watchers to properly disable and enable save button
