+++ app/src/main/res/drawable/divider_green_horizontal.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
+  <solid android:color="@color/ksr_green_500"/>
+  <size android:height="2dp"/>
+</shape>
