+++ app/src/main/res/layout/activity_new_card.xml
@@ -1,114 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto"
-  xmlns:tools="http://schemas.android.com/tools"
-  android:layout_width="match_parent"
+<fragment
+  android:id="@+id/fragment_new_card"
+  android:name="com.kickstarter.ui.fragments.NewCardFragment"
   android:layout_height="match_parent"
-  android:background="@color/ksr_grey_100">
-
-  <android.support.design.widget.AppBarLayout
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content">
-
-    <com.kickstarter.ui.toolbars.KSToolbar
-      android:id="@+id/new_card_toolbar"
-      style="@style/Toolbar"
-      app:contentInsetLeft="0dp"
-      app:contentInsetStart="0dp">
-
-      <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:gravity="center_vertical"
-        android:orientation="horizontal">
-
-        <com.kickstarter.ui.views.IconButton
-          android:id="@+id/back_button"
-          style="@style/ToolbarIconBackButton" />
-
-        <TextView
-          style="@style/ToolbarTitle" />
-      </RelativeLayout>
-
-    </com.kickstarter.ui.toolbars.KSToolbar>
-  </android.support.design.widget.AppBarLayout>
-
-  <ScrollView
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingTop="@dimen/grid_3"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior">
-
-    <FrameLayout
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content">
-
-      <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-        <LinearLayout
-          style="@style/SettingsLinearRow">
-
-          <android.support.design.widget.TextInputLayout
-            style="@style/TextInputLayout"
-            android:layout_marginBottom="0dp"
-            android:layout_marginTop="@dimen/grid_3"
-            app:hintTextAppearance="@style/BodySecondary"
-            app:passwordToggleEnabled="true">
-
-            <android.support.design.widget.TextInputEditText
-              android:id="@+id/cardholder_name"
-              style="@style/ChangeEmailPassTextInputEditText"
-              android:hint="@string/Name"
-              android:inputType="textPersonName"
-              android:maxLines="1" />
-
-          </android.support.design.widget.TextInputLayout>
-
-          <com.stripe.android.view.CardInputWidget
-            android:id="@+id/card_input_widget"
-            style="@style/CardInputWidget"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content" />
-
-          <android.support.design.widget.TextInputLayout
-            style="@style/TextInputLayout"
-            android:layout_marginBottom="0dp"
-            app:hintTextAppearance="@style/BodySecondary"
-            app:passwordToggleEnabled="true">
-
-            <android.support.design.widget.TextInputEditText
-              android:id="@+id/postal_code"
-              style="@style/ChangeEmailPassTextInputEditText"
-              android:hint="@string/Zip_code"
-              android:inputType="phone"
-              android:maxLines="1" />
-
-          </android.support.design.widget.TextInputLayout>
-
-        </LinearLayout>
-
-      </LinearLayout>
-
-      <FrameLayout
-        android:id="@+id/progress_bar"
-        android:layout_width="match_parent"
-        android:layout_height="4dp"
-        android:visibility="gone"
-        tools:visibility="visible">
-
-        <ProgressBar
-          style="?android:attr/progressBarStyleHorizontal"
-          android:layout_width="match_parent"
-          android:layout_height="@dimen/grid_5"
-          android:layout_gravity="center"
-          android:indeterminate="true" />
-
-      </FrameLayout>
-
-    </FrameLayout>
-  </ScrollView>
-
-</android.support.design.widget.CoordinatorLayout>
+  android:layout_width="match_parent"
+  xmlns:android="http://schemas.android.com/apk/res/android" />
