+++ app/src/main/res/layout/fragment_new_card.xml
@@ -1,0 +1,128 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
+  xmlns:app="http://schemas.android.com/apk/res-auto"
+  xmlns:tools="http://schemas.android.com/tools"
+  android:layout_width="match_parent"
+  android:layout_height="match_parent"
+  android:background="@color/ksr_grey_100">
+
+  <android.support.design.widget.AppBarLayout
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content">
+
+    <com.kickstarter.ui.toolbars.KSToolbar
+      android:id="@+id/new_card_toolbar"
+      style="@style/Toolbar"
+      app:contentInsetLeft="0dp"
+      app:contentInsetStart="0dp">
+
+      <RelativeLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:gravity="center_vertical"
+        android:orientation="horizontal">
+
+        <com.kickstarter.ui.views.IconButton
+          android:id="@+id/back_button"
+          style="@style/ToolbarIconBackButton" />
+
+        <TextView
+          style="@style/ToolbarTitle"
+          android:text="@string/Add_new_card"/>
+      </RelativeLayout>
+
+    </com.kickstarter.ui.toolbars.KSToolbar>
+  </android.support.design.widget.AppBarLayout>
+
+  <ScrollView
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    app:layout_behavior="@string/appbar_scrolling_view_behavior">
+
+    <FrameLayout
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content">
+
+      <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:paddingTop="@dimen/grid_3"
+        android:orientation="vertical">
+
+        <LinearLayout
+          style="@style/SettingsLinearRow">
+
+          <android.support.design.widget.TextInputLayout
+            style="@style/TextInputLayout"
+            android:layout_marginBottom="0dp"
+            app:hintTextAppearance="@style/BodySecondary"
+            app:passwordToggleEnabled="true">
+
+            <android.support.design.widget.TextInputEditText
+              android:id="@+id/cardholder_name"
+              style="@style/ChangeEmailPassTextInputEditText"
+              android:hint="@string/Name"
+              android:inputType="textPersonName"
+              android:maxLines="1" />
+
+          </android.support.design.widget.TextInputLayout>
+
+          <com.stripe.android.view.CardInputWidget
+            android:id="@+id/card_input_widget"
+            style="@style/CardInputWidget"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" />
+
+          <FrameLayout
+            android:layout_width="match_parent"
+            android:layout_height="2dp">
+
+            <ImageView
+              android:contentDescription="@null"
+              android:id="@+id/card_focus"
+              android:src="@drawable/divider_dark_grey_500_horizontal"
+              android:layout_marginStart="@dimen/grid_3"
+              android:layout_marginEnd="@dimen/grid_3"
+              android:layout_width="match_parent"
+              android:layout_height="wrap_content"/>
+
+          </FrameLayout>
+
+          <android.support.design.widget.TextInputLayout
+            style="@style/TextInputLayout"
+            app:hintTextAppearance="@style/BodySecondary"
+            app:passwordToggleEnabled="true">
+
+            <android.support.design.widget.TextInputEditText
+              android:id="@+id/postal_code"
+              style="@style/ChangeEmailPassTextInputEditText"
+              android:hint="@string/Zip_code"
+              android:inputType="phone"
+              android:maxLines="1" />
+
+          </android.support.design.widget.TextInputLayout>
+
+        </LinearLayout>
+
+      </LinearLayout>
+
+      <FrameLayout
+        android:id="@+id/progress_bar"
+        android:layout_width="match_parent"
+        android:layout_height="4dp"
+        android:visibility="gone"
+        tools:visibility="visible">
+
+        <ProgressBar
+          style="?android:attr/progressBarStyleHorizontal"
+          android:layout_width="match_parent"
+          android:layout_height="@dimen/grid_5"
+          android:layout_gravity="center"
+          android:indeterminate="true" />
+
+      </FrameLayout>
+
+    </FrameLayout>
+  </ScrollView>
+
+</android.support.design.widget.CoordinatorLayout>
